+++ CHANGELOG.md
@@ -16,6 +16,47 @@ Types of changes
 
 # Latch SDK Changelog
 
+## 2.34.0 - 2023-10-04
+
+### Added
+
+* `directory` modifier for input / outputs
+* Support `temp` by removing from compiled rules. All files / directories are
+temporary because they are deleted at the end of each job on Latch.
+* `multiext` output modifier
+* `report` output modifier
+* `params` in rules
+
+### Fixed
+
+* Replace skipped rules with `Ellipsis`. Supports rules nested in conditionals where previously an empty block was produced.
+* Patched parser to generate compiled code for `workflow.include` calls Compiled workflow.include should carry `print_compilation` keyword (snakemake/snakemake#2469)
+* Detect use of `conda` keyword and install in image. This effectively supports wrapper/conda keywords.
+* `Iterable, Generator` cause issues as type hints when imported from `collections.abc` rather than `typing`
+
+## 2.33.0 - 2023-09-29
+
+### Added
+
+* Add `latch sync` for synchronization from local to remote directories that only uploads modified content
+
+## 2.32.8 - 2023-09-07
+
+### Fixed
+
+* Snakemake:
+    * Better errors if `Snakefile` or `latch_metadata.py` file missing
+    * Correct issues with snakemake example project
+
+## 2.32.7 - 2023-09-07
+
+### Fixed
+
+* Snakemake:
+    * `--snakemake` for `latch dockerfile` command to generate `Dockerfile` with
+    necessary instructions
+    * Snakemake example for `latch init`
+
 ## 2.32.6 - 2023-09-07
 
 ### Fixed